feat: add check if passed id is valid

+++ LabManager/Program.cs
@@ -40,7 +40,8 @@ if (modelName == "Computer")
             Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");
         }
 
-        else {
+        else 
+        {
             Console.WriteLine($"O computador {id} não existe");
         }
         
@@ -49,14 +50,31 @@ if (modelName == "Computer")
     if(modelAction == "Delete")
     {
         //if
-        computerRepository.Delete(Convert.ToInt32(args[2]));
-        Console.WriteLine("Completed computer deletion.");
+        var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
+            computerRepository.Delete(id);
+            Console.WriteLine("Completed computer deletion.");
+        }
+
+         else 
+        {
+            Console.WriteLine($"O computador {id} não existe");
+        }
+       
     }
     if (modelAction == "Update")
     {
-        //update
-        var computer = computerRepository.Update(new Computer(Convert.ToInt32(args[2]), args[3], args[4]));
-        Console.WriteLine($"Updated computer: {computer.Id}, {computer.Ram}, {computer.Processor}");
+        var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
+            var computer = computerRepository.Update(new Computer(Convert.ToInt32(args[2]), args[3], args[4]));
+            Console.WriteLine($"Updated computer: {computer.Id}, {computer.Ram}, {computer.Processor}");
+        }
+        else
+        {
+            Console.WriteLine($"O computador {id} não existe");
+        }
     }
 
 }
@@ -78,20 +96,47 @@ if (modelName == "Lab")
 
     if(modelAction == "Delete")
     {
-        labRepository.Delete(Convert.ToInt32(args[2]));
-        Console.WriteLine("Completed lab deletion.");
+        var id = Convert.ToInt32(args[2]);
+        if(labRepository.ExistsById(id))
+        {
+            labRepository.Delete(Convert.ToInt32(args[2]));
+            Console.WriteLine("Completed lab deletion.");
+        }
+        else
+        {
+            Console.WriteLine($"O Lab {id} não existe");
+        }
+       
     }
 
     if(modelAction == "Show")
     {
-        var lab = labRepository.GetById(Convert.ToInt32(args[2]));
-        Console.WriteLine($"{ lab.Id}, { lab.Number}, {lab.Name}, {lab.Block}");
+        var id = Convert.ToInt32(args[2]);
+        if(labRepository.ExistsById(id))
+        {
+            var lab = labRepository.GetById(Convert.ToInt32(args[2]));
+            Console.WriteLine($"{ lab.Id}, { lab.Number}, {lab.Name}, {lab.Block}");
+        }
+        else
+        {
+            Console.WriteLine($"O Lab {id} não existe");
+        }
+        
     }
 
     if(modelAction == "Update")
     {
-        var lab = labRepository.Update(new Lab(Convert.ToInt32(args[2]), Convert.ToInt32(args[3]), args[4], args[5]));
-        Console.WriteLine($"Updated lab: { lab.Id}, { lab.Number}, {lab.Name}, {lab.Block}");
+        var id = Convert.ToInt32(args[2]);
+        if(labRepository.ExistsById(id))
+        {
+            var lab = labRepository.Update(new Lab(Convert.ToInt32(args[2]), Convert.ToInt32(args[3]), args[4], args[5]));
+            Console.WriteLine($"Updated lab: { lab.Id}, { lab.Number}, {lab.Name}, {lab.Block}");
+        }
+        else
+        {
+            Console.WriteLine($"O Lab {id} não existe");
+        }
+        
     }
 }
